track passwordless logins for groups and users
track passwordless logins for groups and users and display
warnings if they happen. also track if the passwords are not
stored in the shadow file but in the /etc/passwd or /etc/group file
respectively. this also means that potentially sensitive data
can be stored by these commands; maybe in the future we need
an option to sanitize the hashes away from it?