improved command handling
we now bail out upon the first error command and if in debug mode we display
the command executed as well as the stderr output. As we now do popen execution
for the commands via Popen() we temporarily disabled two commands that required
actual shell interpretation as those need to be rewritten and they were not
that important. The find commands with ls executions now don't redirect stderr
anymore to /dev/null as that output is now properly handled.